pyproject.toml
[tool.poetry]
name = "disconcept"
version = "0.1.0"
description = "Automatic Concept Discovery by MST"
authors = ["RenChu Wang <patrick1031wang@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
alive-progress = "^3.1.5"
googlenews = "^1.6.11"
hydra-core = "^1.3.2"
loguru = "^0.7.2"
newspaper3k = "^0.2.8"
nltk = "^3.8.1"
numpy = "^1.26.2"
python = "^3.10"
scipy = "^1.11.4"
torch = "^2.1.1"
torch-geometric = "^2.4.0"
wikipedia = "^1.4.0"


[tool.poetry.group.dev.dependencies]
autoflake = "^2.2.1"
black = "^23.11.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
